# Conflicts:
#   src/main/java/com/kh/hotels/mngStock/controller/StockController.java
#   src/main/java/com/kh/hotels/mngStock/model/dao/StockDaoImpl.java
#   src/main/resources/mappers/stock-mapper.xml
#   src/main/webapp/WEB-INF/views/hoteladmin/mngStock/stock/repairRequest.jsp

+++ src/main/java/com/kh/hotels/mngApproval/model/service/ApprovalService.java
@@ -76,7 +76,7 @@ public interface ApprovalService {
 
 	ArrayList<HashMap<String, Object>> selectOrderDocu();
 
-	ArrayList<HashMap<String, Object>> selectOrderList(int dnum);
+	ArrayList<HashMap<String, Object>> selectOrderList(Long dnum);
 
 	int insertOrderList(ArrayList<OrderRequest> oRequestList);
